Protected goal routes

+++ backend/controllers/goalController.js
@@ -1,11 +1,12 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 // @desc    Get goals
 // @route   GET api/goals
 // @access  Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
 
   res.status(200).json(goals);
 });
@@ -20,6 +21,7 @@ const setGoal = asyncHandler(async (req, res) => {
   }
 
   const goal = await Goal.create({
+    user: req.user.id,
     text: req.body.text,
   });
   res.status(200).json(goal);
@@ -33,6 +35,14 @@ const updateGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found!");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  // Make sure the loggend in user matches goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("Not authorized");
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -51,6 +61,14 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found!");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  // Make sure the loggend in user matches goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("Not authorized");
   }
 
   await goal.deleteOne();
